refactor(ui): remove global FormError component
