--- packages/polymath-ui/src/components/FormItemGroup/GroupError/index.js
@@ -1,22 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import { ErrorMessage } from 'formik';
-
-import { formError } from '../../../styles/utils';
-
-const StyledError = styled.span`
-  ${formError};
-  margin-top: 0.25rem;
-  float: left;
-  margin-right: 0.5rem;
-
-  &:last-child {
-    margin-right: 0;
-  }
-`;
-
-const FormError = ({ name }) => (
-  <ErrorMessage component={StyledError} name={name} />
-);
-
-export default FormError;
