+++ packages/polymath-ui/src/index.js
@@ -11,7 +11,6 @@ export { default as Heading } from './components/Heading';
 export { default as Paragraph } from './components/Paragraph';
 export { default as FormItem } from './components/FormItem';
 export { default as FormItemGroup } from './components/FormItemGroup';
-export { default as FormError } from './components/FormError';
 export { default as Modal } from './components/Modal';
 export { default as STOStatus } from './components/STOStatus';
 export { default as Countdown } from './components/Countdown';
